Extract UnsafeAccessor generation code.
This extracts the UnsafeAccessor generator code
and places it in its compilation unit.
